+++ modules/gcp_archival_location/outputs.tf
@@ -1,0 +1,4 @@
+output "archival_location_id" {
+  description = "RSC archival location ID."
+  value       = polaris_gcp_archival_location.archival_location.id
+}