+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
 
 ## v1.0.0
-*  Add `gcp_project` Terraform module.
+* Add `gcp_project` Terraform module.
+* Add `gcp_archival_location` Terraform module.